[level 1] Title: 크기가 작은 부분문자열, Time: 0.16 ms, Memory: 78.8 MB -BaekjoonHub

+++ 프로그래머스/1/147355. 크기가 작은 부분문자열/크기가 작은 부분문자열.java
@@ -1,6 +1,6 @@
 class Solution {
-    public Long solution(String t, String p) {
-        Long answer = 0l;
+    public int solution(String t, String p) {
+        int answer = 0;
         for(int i = 0; i < t.length() - p.length() + 1; i++){
             Long num = Long.parseLong(t.substring(i,i+p.length()));
             if(num <= Long.parseLong(p)) answer++;
